Renamed ZipExtractor to Installer Remove existing folders when upgrading
